First draft and moving loading and saving in BaseModule
