+++ objects/models/__init__.py
@@ -2,7 +2,6 @@ from objects.models.transformer import Transformer, TransformerXL
 from objects.models.memory import EntNet, NPN #, ProStruct
 from objects.models.glad import GLAD
 from objects.models.seq2seq import Seq2Seq
-from objects.models.agent_dqn import AgentDQN
-from objects.models.nlu_nlg import NLU, NLG
+from objects.models.ddq import DQN, NLU, NLG
 from objects.models.policies import *
