+++ objects/modules/__init__.py
@@ -1,7 +1,7 @@
-from objects.modules.belief_tracker import *
-from objects.modules.policy_manager import *
-from objects.modules.text_generator import *
+from objects.modules.belief_tracker import RuleBeliefTracker, NeuralBeliefTracker
+from objects.modules.policy_manager import RulePolicyManager, NeuralPolicyManager
+from objects.modules.text_generator import RuleTextGenerator, NeuralTextGenerator
 from objects.modules.kb_operator import KBHelper
-from objects.modules.agent import DialogManager
-from objects.modules.user import RuleSimulator, ModelBasedSimulator
+from objects.modules.dialog_manager import DialogManager
+from objects.modules.user import RuleSimulator, NeuralSimulator
