+++ objects/modules/belief_tracker.py
@@ -2,7 +2,9 @@ import numpy as np
 import os, pdb, sys  # set_trace
 
 from torch import nn
+from objects.blocks.base import BaseBeliefTracker
 
+'''
 class BaseBeliefTracker(object):
   def __init__(self, data):
     self.data = data
@@ -12,12 +14,11 @@ class BaseBeliefTracker(object):
     raise NotImplementedError
 
   def predict(self):
-    '''
     a belief tracker predicts user intents:
       input - user utterance, previous embedded context, memory
       output - a binary value for each possible slot in the ontology
-    '''
     raise NotImplementedError
+'''
 
 
 class RuleBeliefTracker(BaseBeliefTracker):
@@ -37,6 +38,30 @@ class RuleBeliefTracker(BaseBeliefTracker):
     input_text
 
 
-class NeuralBeliefTracker(BaseBeliefTracker, nn.Module):
-  def __init__(self):
-    super().__init__()
+class NeuralBeliefTracker(BaseBeliefTracker):
+  def __init__(self, args, model):
+    super().__init__(args, model)
+
+  def extract_predictions(self, scores, threshold=0.5):
+    batch_size = len(list(scores.values())[0])
+    predictions = [set() for i in range(batch_size)]
+    for s in self.ontology.slots:
+      for i, p in enumerate(scores[s]):
+        triggered = [(s, v, p_v) for v, p_v in zip(self.ontology.values[s], p) if p_v > threshold]
+        if s == 'request':
+          # we can have multiple requests predictions
+          predictions[i] |= set([(s, v) for s, v, p_v in triggered])
+        elif triggered:
+          # only extract the top inform prediction
+          sort = sorted(triggered, key=lambda tup: tup[-1], reverse=True)
+          predictions[i].add((sort[0][0], sort[0][1]))
+    return predictions
+
+  def run_glad_inference(self, data):
+    self.eval()
+    predictions = []
+    for batch in data.batch(self.batch_size):
+      loss, scores = self.forward(batch)
+      predictions += self.extract_predictions(scores)
+    return predictions
+
